+++ app/layout.js
@@ -4,8 +4,8 @@ import "./globals.css";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Tarea frontend",
+  description: "Es una aplicación de frontend",
 };
 
 export default function RootLayout({ children }) {
