+++ app/users/page.js
@@ -1,8 +1,6 @@
 "use client"
 import React, {useEffect, useState} from 'react';
-import {Container, Table, TableBody, TableCell, TableHead, TableRow} from "@mui/material";
-import IconButton from '@mui/material/IconButton';
-import {Edit} from "@mui/icons-material";
+import {Container} from "@mui/material";
 
 import AuthService from "../../services/AuthService";
 import { useRouter } from 'next/navigation';
@@ -37,42 +35,10 @@ export default function Users(){
         setUsers([data]);
     }
 
-    const handleEdit = (user) => {
-        router.push('/users/' + user.id + '/edit');
-    }
-
     return (
         <Container>
             <Navbar />
-            <h1>Users</h1>
-            <Table>
-                <TableHead>
-                    <TableRow>
-                        <TableCell>Nombre</TableCell>
-                        <TableCell>Email</TableCell>
-                        <TableCell>Estado</TableCell>
-                        <TableCell>Última Sesión</TableCell>
-                        <TableCell>Acciones</TableCell>
-                    </TableRow>
-                </TableHead>
-                <TableBody>
-                    {
-                        users.map((user) => (
-                            <TableRow key={user}>
-                                <TableCell>{user.name}</TableCell>
-                                <TableCell>{user.email}</TableCell>
-                                <TableCell>{user.status? 'ACTIVO' : 'CERRADO'}</TableCell>
-                                <TableCell>TBD</TableCell>
-                                <TableCell>
-                                    <IconButton color="primary" aria-label={"Editar usuario " + user.name} onClick={() => handleEdit(user)}>
-                                        <Edit />
-                                    </IconButton>
-                                </TableCell>
-                            </TableRow>
-                        ))
-                    }
-                </TableBody>
-            </Table>
+            <h1>Inicio</h1>
         </Container>
     )
 }
