+++ components/Navbar.css
@@ -1,7 +1,6 @@
 @import '../app/globals.css';
 
 .navbar {
-    display: -webkit-flex;
     display: flex;
     -webkit-flex-direction: row;
     flex-direction: row;
