+++ components/Navbar.js
@@ -1,6 +1,7 @@
 import React, {useEffect, useState} from 'react';
 import './Navbar.css';
 import Button from '@mui/material/Button';
+import { Stack } from '@mui/material';
 import AuthService from '@/services/AuthService';
 import { useRouter } from 'next/navigation';
 
@@ -10,6 +11,7 @@ const Navbar = () => {
 
     useEffect(() => {
         setUser(JSON.parse(localStorage.getItem('user')));
+        
     }, []);
 
     const handleLogout = async () => {
@@ -19,17 +21,38 @@ const Navbar = () => {
             router.push('/login');
         }
     }
+    const handleNavigatorHome= () => {
+        router.push('/users/');
+    }
+
+    const handleNavigator= (route) => {
+        router.push('/users/' + user.id + '/'+ route);
+    }
 
     return (
         <div className="navbar">
             <div className="navbar-item">
                 {user?.name}
             </div>
-            <div className="navbar-item">
-                <Button onClick={handleLogout}>
+            <Button variant='outlined' onClick={() => handleNavigatorHome()} >
+                Inicio
+            </Button>
+            <Stack direction='row' spacing={2}>
+            
+            <Button variant='outlined' onClick={() => handleNavigator('search')} >
+                Buscador
+            </Button>
+            <Button variant='outlined' onClick={() => handleNavigator('bulkCreate')} >
+                Creador de usuarios
+            </Button>
+            <Button variant='outlined' onClick={() => handleNavigator('edit')}>
+                Editor de usuario
+            </Button>
+            </Stack>
+                <Button onClick={handleLogout} className='button'>
                     Logout
                 </Button>
-            </div>
+            
         </div>
     )
 }
